CloudManagerService refactoring. Added GeneralUplinkMessageService and TsUplinkMessageService
