+++ application/src/main/java/org/thingsboard/server/service/cloud/GeneralUplinkMessageService.java
@@ -1,0 +1,28 @@
+/**
+ * Copyright © 2016-2024 The Thingsboard Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.thingsboard.server.service.cloud;
+
+import com.google.common.util.concurrent.ListenableFuture;
+import org.thingsboard.server.common.data.id.TenantId;
+import org.thingsboard.server.common.data.page.TimePageLink;
+
+public interface GeneralUplinkMessageService extends UplinkMessageService {
+
+    TimePageLink newCloudEventsAvailable(TenantId tenantId, Long queueSeqIdStart);
+
+    ListenableFuture<Long> getQueueStartTs(TenantId tenantId);
+}
+
