+++ msa/edge-black-box-tests/src/test/java/org/thingsboard/server/msa/AbstractContainerTest.java
@@ -38,6 +38,7 @@ import org.thingsboard.server.common.data.DeviceProfileType;
 import org.thingsboard.server.common.data.DeviceTransportType;
 import org.thingsboard.server.common.data.EntityType;
 import org.thingsboard.server.common.data.EntityView;
+import org.thingsboard.server.common.data.HasVersion;
 import org.thingsboard.server.common.data.OtaPackage;
 import org.thingsboard.server.common.data.OtaPackageInfo;
 import org.thingsboard.server.common.data.StringUtils;
@@ -186,7 +187,12 @@ public abstract class AbstractContainerTest {
                 .atMost(30, TimeUnit.SECONDS).
                 until(() -> {
                     try {
-                        return edgeRestClient.getWidgetsBundles(new PageLink(100)).getTotalElements() == 26;
+                        long totalElements = edgeRestClient.getWidgetsBundles(new PageLink(100)).getTotalElements();
+                        final long expectedCount = 28;
+                        if (totalElements != expectedCount) {
+                            log.warn("Expected {} widget bundles, but got {}", expectedCount, totalElements);
+                        }
+                        return totalElements == expectedCount;
                     } catch (Throwable e) {
                         return false;
                     }
@@ -447,6 +453,7 @@ public abstract class AbstractContainerTest {
             expected.setDefaultRuleChainId(null);
             actual.setDefaultEdgeRuleChainId(null);
             actual.setDefaultRuleChainId(null);
+            cleanUpVersion(expected, actual);
             Assert.assertEquals("Device profiles on cloud and edge are different", expected, actual);
         }
     }
@@ -461,6 +468,7 @@ public abstract class AbstractContainerTest {
             Assert.assertEquals(expected.getDefaultRuleChainId(), actual.getDefaultEdgeRuleChainId());
             expected.setDefaultRuleChainId(null);
             actual.setDefaultEdgeRuleChainId(null);
+            cleanUpVersion(expected, actual);
             Assert.assertEquals("Asset profiles on cloud and edge are different", expected, actual);
         }
     }
@@ -494,6 +502,7 @@ public abstract class AbstractContainerTest {
             Assert.assertEquals("Cloud rule chain type is incorrect", RuleChainType.EDGE, actual.getType());
             expected.setType(null);
             actual.setType(null);
+            cleanUpVersion(expected, actual);
             Assert.assertEquals("Rule chains on cloud and edge are different (except type)", expected, actual);
 
             Awaitility.await()
@@ -546,6 +555,7 @@ public abstract class AbstractContainerTest {
             WidgetsBundle actual = cloudWidgetsBundle.get();
             expected.setImage(null);
             actual.setImage(null);
+            cleanUpVersion(expected, actual);
             Assert.assertEquals("Widgets bundles on cloud and edge are different", expected, actual);
         }
     }
@@ -559,6 +569,7 @@ public abstract class AbstractContainerTest {
             WidgetTypeDetails actual = cloudWidgetsBundle.get();
             expected.setImage(null);
             actual.setImage(null);
+            cleanUpVersion(expected, actual);
             Assert.assertEquals("Widget types on cloud and edge are different", expected, actual);
         }
     }
@@ -570,6 +581,7 @@ public abstract class AbstractContainerTest {
             Optional<Device> cloudDevice = cloudRestClient.getDeviceById(deviceId);
             Device expected = edgeDevice.get();
             Device actual = cloudDevice.get();
+            cleanUpVersion(expected, actual);
             Assert.assertEquals("Devices on cloud and edge are different", expected, actual);
         }
     }
@@ -581,6 +593,7 @@ public abstract class AbstractContainerTest {
             Optional<Asset> cloudAsset = cloudRestClient.getAssetById(assetId);
             Asset expected = edgeAsset.get();
             Asset actual = cloudAsset.get();
+            cleanUpVersion(expected, actual);
             Assert.assertEquals("Assets on cloud and edge are different", expected, actual);
         }
     }
@@ -592,6 +605,7 @@ public abstract class AbstractContainerTest {
             Optional<EntityView> cloudEntityView = cloudRestClient.getEntityViewById(entityViewId);
             EntityView expected = edgeEntityView.get();
             EntityView actual = cloudEntityView.get();
+            cleanUpVersion(expected, actual);
             Assert.assertEquals("Entity Views on cloud and edge are different", expected, actual);
         }
     }
@@ -603,6 +617,7 @@ public abstract class AbstractContainerTest {
             Optional<Dashboard> cloudDashboard = cloudRestClient.getDashboardById(dashboardId);
             Dashboard expected = edgeDashboard.get();
             Dashboard actual = cloudDashboard.get();
+            cleanUpVersion(expected, actual);
             Assert.assertEquals("Dashboards on cloud and edge are different", expected, actual);
         }
     }
@@ -616,8 +631,14 @@ public abstract class AbstractContainerTest {
             User actual = cloudUser.get();
             expected.setAdditionalInfo(cleanLastLoginTsFromAdditionalInfo(expected.getAdditionalInfo()));
             actual.setAdditionalInfo(cleanLastLoginTsFromAdditionalInfo(actual.getAdditionalInfo()));
+            cleanUpVersion(expected, actual);
             Assert.assertEquals("Users on cloud and edge are different (except lastLoginTs)", expected, actual);
         }
+    }
+
+    protected void cleanUpVersion(HasVersion expected, HasVersion actual) {
+        expected.setVersion(null);
+        actual.setVersion(null);
     }
 
     private JsonNode cleanLastLoginTsFromAdditionalInfo(JsonNode additionalInfo) {
